Updated the PUT method.
Attempts to add information into the URL were futile.

Too allow for maximum flexibility as well as efficient handling of the
data, making use of the request bod ywas mandatory.

This has been testted to work with updating through Postman as well as
updating from the website istelf when lessons are bought.